Run prettier again; adjust bolding; no select OLs
The problem with li:first-line as a way to bold the titles of list
entries was that, on small-width screens, long titles (like the
thesis') would span multiple lines, and only the stuff on the
first line would be bolded!

A simple solution is just using <span>s to indicate what stuff
counts as the "name" of a list entry, and should thus be bolded.
bit more verbose but it works fine